// apps/scanner/src/features/fill-container/model/machine/fill-container.effects.ts
import { BucketQRData, IFillingActBucketResponseDto } from '@repo/api';

export type FillContainerEffect =
  | { type: 'SHOW_SCANNER_PROGRESS' }
  | { type: 'VALIDATE_BUCKET'; qrCode: string }
  | { type: 'SHOW_BUCKET_SUCCESS'; bucket: BucketQRData }
  | { type: 'SHOW_BUCKET_ERROR'; message: string }
  | { type: 'VALIDATE_COMPONENT'; barCode: string; bucket: BucketQRData }
  | {
      type: 'SHOW_COMPONENT_SUCCESS';
      fillingAct: IFillingActBucketResponseDto;
      bucket: BucketQRData;
    }
  | { type: 'SHOW_COMPONENT_ERROR'; message: string };